Added parameter to look_room. Fixes #9.
Adding a parameter is not awesome, but passing the exit object into the function is a clean way to check if the exit in the look direction is locked. The alternative would be to do a check before calling look_room but that would need to be added on each exit. Also this way the message is consistant.